+++ Jvedio-WPF/Jvedio/Utils/Common/Converter.cs
@@ -42,7 +42,7 @@ namespace Jvedio
         // 数字转换为选中项的地址
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            if (value == null | parameter == null) {
+            if (value == null || parameter == null) {
                 return false;
             }
 
@@ -56,7 +56,7 @@ namespace Jvedio
         // 选中项地址转换为数字
         public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            if (value == null | parameter == null)
+            if (value == null || parameter == null)
                 return 0;
             int intparameter = int.Parse(parameter.ToString());
             return (VideoType)intparameter;